+++ core/src/game_state/mod.rs
@@ -1,19 +1,18 @@
 use std::ops::Deref;
 
-use anyhow::{anyhow, bail, Error};
+use anyhow::{bail, Error};
 use serde::{Deserialize, Serialize};
 
-use crate::bidding::Bid;
-use crate::deck::Deck;
-use crate::hands::Hands;
 use crate::message::MessageVariant;
-use crate::settings::{FirstLandlordSelectionPolicy, GameMode, KittyBidPolicy, PropagatedState};
-use crate::types::{Card, PlayerID, Rank, Trump};
+use crate::settings::PropagatedState;
+use crate::types::PlayerID;
 
+pub mod draw_phase;
 pub mod exchange_phase;
 pub mod initialize_phase;
 pub mod play_phase;
 
+use draw_phase::DrawPhase;
 use exchange_phase::ExchangePhase;
 use initialize_phase::InitializePhase;
 use play_phase::PlayPhase;
@@ -139,20 +138,8 @@ impl GameState {
         let mut s = self.clone();
         match s {
             GameState::Initialize { .. } => (),
-            GameState::Draw(DrawPhase {
-                ref mut hands,
-                ref mut kitty,
-                ref mut deck,
-                revealed_cards,
-                ..
-            }) => {
-                hands.destructively_redact_except_for_player(id);
-                for card in &mut kitty[revealed_cards..] {
-                    *card = Card::Unknown;
-                }
-                for card in deck {
-                    *card = Card::Unknown;
-                }
+            GameState::Draw(ref mut p) => {
+                p.destructively_redact_for_player(id);
             }
             GameState::Exchange(ref mut p) => {
                 p.destructively_redact_for_player(id);
@@ -170,263 +157,6 @@ impl Deref for GameState {
 
     fn deref(&self) -> &PropagatedState {
         self.propagated()
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DrawPhase {
-    num_decks: usize,
-    game_mode: GameMode,
-    deck: Vec<Card>,
-    propagated: PropagatedState,
-    hands: Hands,
-    bids: Vec<Bid>,
-    #[serde(default)]
-    autobid: Option<Bid>,
-    position: usize,
-    kitty: Vec<Card>,
-    #[serde(default)]
-    revealed_cards: usize,
-    level: Option<Rank>,
-    #[serde(default)]
-    removed_cards: Vec<Card>,
-    #[serde(default)]
-    decks: Vec<Deck>,
-}
-
-impl DrawPhase {
-    pub fn propagated(&self) -> &PropagatedState {
-        &self.propagated
-    }
-
-    pub fn propagated_mut(&mut self) -> &mut PropagatedState {
-        &mut self.propagated
-    }
-
-    pub fn add_observer(&mut self, name: String) -> Result<PlayerID, Error> {
-        self.propagated.add_observer(name)
-    }
-
-    pub fn remove_observer(&mut self, id: PlayerID) -> Result<(), Error> {
-        self.propagated.remove_observer(id)
-    }
-
-    pub fn next_player(&self) -> Result<PlayerID, Error> {
-        if self.deck.is_empty() {
-            let (first_bid, winning_bid) = Bid::first_and_winner(&self.bids, self.autobid)?;
-            let landlord = self.propagated.landlord.unwrap_or(
-                match self.propagated.first_landlord_selection_policy {
-                    FirstLandlordSelectionPolicy::ByWinningBid => winning_bid.id,
-                    FirstLandlordSelectionPolicy::ByFirstBid => first_bid.id,
-                },
-            );
-
-            Ok(landlord)
-        } else {
-            Ok(self.propagated.players[self.position].id)
-        }
-    }
-
-    pub fn draw_card(&mut self, id: PlayerID) -> Result<(), Error> {
-        if id != self.propagated.players[self.position].id {
-            bail!("not your turn!");
-        }
-        if let Some(next_card) = self.deck.pop() {
-            self.hands.add(id, Some(next_card))?;
-            self.position = (self.position + 1) % self.propagated.players.len();
-            Ok(())
-        } else {
-            bail!("no cards left in deck")
-        }
-    }
-
-    pub fn reveal_card(&mut self) -> Result<MessageVariant, Error> {
-        if !self.deck.is_empty() {
-            bail!("can't reveal card until deck is fully drawn")
-        }
-        if !self.bids.is_empty() {
-            bail!("can't reveal card if at least one bid has been made")
-        }
-        let id = self
-            .propagated
-            .landlord
-            .ok_or_else(|| anyhow!("can't reveal card if landlord hasn't been selected yet"))?;
-
-        let landlord_level = self
-            .propagated
-            .players
-            .iter()
-            .find(|p| p.id == id)
-            .ok_or_else(|| anyhow!("Couldn't find landlord level?"))?
-            .rank();
-
-        if landlord_level == Rank::NoTrump {
-            bail!("can't reveal card if the level is no trump!");
-        }
-
-        if self.revealed_cards >= self.kitty.len() || self.autobid.is_some() {
-            bail!("can't reveal any more cards")
-        }
-
-        let level = self
-            .propagated
-            .players
-            .iter()
-            .find(|p| p.id == id)
-            .map(|p| p.rank())
-            .ok_or_else(|| anyhow!("can't find landlord level?"))?;
-
-        let card = self.kitty[self.revealed_cards];
-
-        match self.propagated.kitty_bid_policy {
-            KittyBidPolicy::FirstCard => {
-                self.autobid = Some(Bid {
-                    count: 1,
-                    id,
-                    card,
-                    epoch: 0,
-                });
-            }
-            KittyBidPolicy::FirstCardOfLevelOrHighest
-                if card.is_joker() || card.number().map(Rank::Number) == Some(level) =>
-            {
-                self.autobid = Some(Bid {
-                    count: 1,
-                    id,
-                    card,
-                    epoch: 0,
-                });
-            }
-            KittyBidPolicy::FirstCardOfLevelOrHighest
-                if self.revealed_cards >= self.kitty.len() - 1 =>
-            {
-                let mut sorted_kitty = self.kitty.clone();
-                sorted_kitty.sort_by(|a, b| {
-                    Trump::NoTrump {
-                        number: match level {
-                            Rank::Number(n) => Some(n),
-                            Rank::NoTrump => None,
-                        },
-                    }
-                    .compare(*a, *b)
-                });
-                if let Some(highest_card) = sorted_kitty.last() {
-                    self.autobid = Some(Bid {
-                        count: 1,
-                        id,
-                        card: *highest_card,
-                        epoch: 0,
-                    });
-                }
-            }
-            _ => (),
-        }
-        self.revealed_cards += 1;
-
-        Ok(MessageVariant::RevealedCardFromKitty)
-    }
-
-    pub fn bid(&mut self, id: PlayerID, card: Card, count: usize) -> bool {
-        if self.revealed_cards > 0 {
-            return false;
-        }
-        Bid::bid(
-            id,
-            card,
-            count,
-            &mut self.bids,
-            self.autobid,
-            &self.hands,
-            &self.propagated.players,
-            self.propagated.landlord,
-            self.propagated.bid_policy,
-            self.propagated.bid_reinforcement_policy,
-            self.propagated.joker_bid_policy,
-            self.num_decks,
-            0,
-        )
-    }
-
-    pub fn take_back_bid(&mut self, id: PlayerID) -> Result<(), Error> {
-        Bid::take_back_bid(id, self.propagated.bid_takeback_policy, &mut self.bids, 0)
-    }
-
-    pub fn done_drawing(&self) -> bool {
-        self.deck.is_empty()
-    }
-
-    pub fn advance(&self, id: PlayerID) -> Result<ExchangePhase, Error> {
-        if !self.deck.is_empty() {
-            bail!("deck has cards remaining")
-        }
-
-        let (landlord, landlord_level) = {
-            let landlord = match self.propagated.landlord {
-                Some(landlord) => landlord,
-                None => {
-                    let (first_bid, winning_bid) = Bid::first_and_winner(&self.bids, self.autobid)?;
-                    match self.propagated.first_landlord_selection_policy {
-                        FirstLandlordSelectionPolicy::ByWinningBid => winning_bid.id,
-                        FirstLandlordSelectionPolicy::ByFirstBid => first_bid.id,
-                    }
-                }
-            };
-
-            if id != landlord {
-                bail!("only the leader can advance the game");
-            }
-            let landlord_level = self
-                .propagated
-                .players
-                .iter()
-                .find(|p| p.id == landlord)
-                .ok_or_else(|| anyhow!("Couldn't find landlord level?"))?
-                .rank();
-            (landlord, landlord_level)
-        };
-        let trump = match landlord_level {
-            Rank::NoTrump => Trump::NoTrump { number: None },
-            Rank::Number(landlord_level) => {
-                // Note: this is not repeated in all cases above, but it is
-                // repeated in some. It's OK because the bid calculation is
-                // fast.
-                let (_, winning_bid) = Bid::first_and_winner(&self.bids, self.autobid)?;
-                match winning_bid.card {
-                    Card::Unknown => bail!("can't bid with unknown cards!"),
-                    Card::SmallJoker | Card::BigJoker => Trump::NoTrump {
-                        number: Some(landlord_level),
-                    },
-                    Card::Suited { suit, .. } => Trump::Standard {
-                        suit,
-                        number: landlord_level,
-                    },
-                }
-            }
-        };
-        let mut hands = self.hands.clone();
-        hands.set_trump(trump);
-        Ok(ExchangePhase::new(
-            self.propagated.clone(),
-            self.num_decks,
-            self.game_mode.clone(),
-            self.kitty.clone(),
-            landlord,
-            hands,
-            trump,
-            self.bids.clone(),
-            self.autobid,
-            self.removed_cards.clone(),
-            self.decks.clone(),
-        ))
-    }
-
-    pub fn return_to_initialize(&self) -> Result<(InitializePhase, Vec<MessageVariant>), Error> {
-        let mut msgs = vec![MessageVariant::ResettingGame];
-
-        let mut propagated = self.propagated.clone();
-        msgs.extend(propagated.make_all_observers_into_players()?);
-
-        Ok((InitializePhase::from_propagated(propagated), msgs))
     }
 }
 
@@ -862,10 +592,10 @@ mod tests {
                     init_.set_num_decks(Some(n_decks)).unwrap();
                     if init_.set_kitty_size(Some(kitty_size)).is_ok() {
                         let draw = init_.start(p1).unwrap();
-                        assert_eq!(draw.deck.len() % n_players, 0);
-                        assert_eq!(draw.kitty.len(), kitty_size);
+                        assert_eq!(draw.deck().len() % n_players, 0);
+                        assert_eq!(draw.kitty().len(), kitty_size);
                         assert_eq!(
-                            draw.removed_cards.len() + draw.deck.len() + draw.kitty.len(),
+                            draw.removed_cards().len() + draw.deck().len() + draw.kitty().len(),
                             n_decks * FULL_DECK.len()
                         );
                     }
@@ -883,7 +613,7 @@ mod tests {
         let p4 = init.add_player("p4".into()).unwrap().0;
         let mut draw = init.start(PlayerID(0)).unwrap();
         // Hackily ensure that everyone can bid.
-        draw.deck = vec![
+        *draw.deck_mut() = vec![
             cards::S_2,
             Card::SmallJoker,
             Card::BigJoker,
@@ -893,7 +623,7 @@ mod tests {
             Card::BigJoker,
             cards::H_2,
         ];
-        draw.position = 0;
+        *draw.position_mut() = 0;
 
         draw.draw_card(p1).unwrap();
         draw.draw_card(p2).unwrap();
@@ -922,7 +652,7 @@ mod tests {
             .unwrap();
         let mut draw = init.start(PlayerID(0)).unwrap();
         // Hackily ensure that everyone can bid.
-        draw.deck = vec![
+        *draw.deck_mut() = vec![
             cards::S_2,
             Card::SmallJoker,
             Card::BigJoker,
@@ -932,7 +662,7 @@ mod tests {
             Card::BigJoker,
             cards::H_2,
         ];
-        draw.position = 0;
+        *draw.position_mut() = 0;
 
         draw.draw_card(p1).unwrap();
         draw.draw_card(p2).unwrap();
@@ -984,8 +714,8 @@ mod tests {
             deck.push(p4_hand[i]);
         }
         deck.reverse();
-        draw.deck = deck;
-        draw.position = 0;
+        *draw.deck_mut() = deck;
+        *draw.position_mut() = 0;
 
         for _ in 0..11 {
             draw.draw_card(p1).unwrap();
@@ -1023,7 +753,7 @@ mod tests {
                 .unwrap();
 
             let mut draw = init.start(PlayerID(1)).unwrap();
-            draw.deck = vec![bid, bid, bid, bid];
+            *draw.deck_mut() = vec![bid, bid, bid, bid];
             draw.draw_card(p2).unwrap();
             draw.draw_card(p3).unwrap();
             draw.draw_card(p4).unwrap();
@@ -1268,8 +998,8 @@ mod tests {
             deck.push(p6_hand[i]);
         }
         deck.reverse();
-        draw.deck = deck;
-        draw.position = 0;
+        *draw.deck_mut() = deck;
+        *draw.position_mut() = 0;
 
         for _ in 0..26 {
             draw.draw_card(p1).unwrap();
@@ -1280,7 +1010,7 @@ mod tests {
             draw.draw_card(p6).unwrap();
         }
 
-        draw.kitty = vec![C_7, S_9, D_6, D_J, C_Q, C_10];
+        *draw.kitty_mut() = vec![C_7, S_9, D_6, D_J, C_Q, C_10];
 
         assert!(draw.bid(p1, D_7, 2));
 
@@ -1501,8 +1231,8 @@ mod tests {
             deck.push(p8_hand[i]);
         }
         deck.reverse();
-        draw.deck = deck;
-        draw.position = 0;
+        *draw.deck_mut() = deck;
+        *draw.position_mut() = 0;
 
         // Draw the deck
         for _ in 0..2 {
